#24 start migration to cesapp

+++ containers/install-containers.sh
@@ -11,21 +11,29 @@ CONTAINERS=$(cat $CONTAINERFILE)
 for C in $CONTAINERS; do
   DIR="$BASEDIR/$C"
 
-  if [ ! -f "$DIR/build.sh" ]; then
-    echo "could not find build script for container $C"
-    exit 1
-  fi
+  if [ -f "$DIR/docker/dogu.json" ]; then
 
-  echo "execute build script for container $C"
-  $DIR/build.sh
-  if [ -f "$DIR/create.sh" ]; then
-    echo "create container $C"
-    $DIR/create.sh
-  fi
+    /opt/ces/bin/cesapp build "$DIR/docker"
 
-  if [ -d "$DIR/resources" ]; then
-    echo "copy container required resources to host system"
-    rsync -rav $DIR/resources/* /
+  else
+
+    if [ ! -f "$DIR/build.sh" ]; then
+      echo "could not find build script for container $C"
+      exit 1
+    fi
+
+    echo "execute build script for container $C"
+    $DIR/build.sh
+    if [ -f "$DIR/create.sh" ]; then
+      echo "create container $C"
+      $DIR/create.sh
+    fi
+
+    if [ -d "$DIR/resources" ]; then
+      echo "copy container required resources to host system"
+      rsync -rav $DIR/resources/* /
+    fi
+
   fi
 
   if [ "$2" = "start" ]; then
